backend/app/requirements.txt
fastapi==0.104.1
supabase==2.0.2
google-generativeai==0.3.2
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2 